requirements.txt
pipwin
pyaudio
SpeechRecognition
openai
python-dotenv
asyncio
edge-tts
playsound==1.2.2
wave
scipy